move: 수평선 파일 이동
